# Remove workaround for astropy bug fixed in astropy 4.3