Fixes https://github.com/cta-observatory/ctapipe/issues/1552