@maxnoe 2023-01-09
So, turns out, with our quickfix, a camera geometry read from file is faster after running to table then before:

```python
import numpy as np
from ctapipe.instrument import SubarrayDescription

s = SubarrayDescription.from_hdf('../public_dl2/data/gamma-diffuse.dl2.h5')
cam = s.tel[1].camera.geometry

pix = np.random.choice(np.arange(len(cam)), 50)

print("Before to_table")
get_ipython().run_line_magic('timeit', 'cam[pix]')

t = cam.to_table()
print("After to_table")
get_ipython().run_line_magic('timeit', 'cam[pix]')
```


On master:
```
Before to_table
27.2 µs ± 2.59 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
After to_table
20.9 µs ± 1.33 µs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

Here:
```
Before to_table
27.9 µs ± 670 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
After to_table
26.4 µs ± 726 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
```

This means to me, instead of removing this workaround, we should probably move it to `CameraGeometry.__init__`.

Astropy deepcopying these meta info objects in all operations is really annoying for performance, I opened a couple of issues, some got fixed already and are pending a new release.